Refactor: 커뮤니티 api파일 폴더 이동
